requirements.txt
flask
pillow
google-genai